+++ dialects/postgresql/src/main/kotlin/app/cash/sqldelight/dialects/postgresql/PostgreSqlType.kt
@@ -32,6 +32,7 @@ enum class PostgreSqlType(override val javaType: TypeName) : DialectType {
   ENUM(STRING),
   GEOMETRY(STRING),
   GEOGRAPHY(STRING),
+  LTREE(STRING),
   ;
 
   override fun prepareStatementBinder(columnIndex: CodeBlock, value: CodeBlock): CodeBlock {
@@ -46,7 +47,7 @@ enum class PostgreSqlType(override val javaType: TypeName) : DialectType {
         value,
       )
       NUMERIC -> CodeBlock.of("bindBigDecimal(%L, %L)\n", columnIndex, value)
-      ENUM, INTERVAL, JSON, TSVECTOR, TSTZRANGE, TSRANGE, TSMULTIRANGE, TSTZMULTIRANGE, TSQUERY -> CodeBlock.of(
+      LTREE, ENUM, INTERVAL, JSON, TSVECTOR, TSTZRANGE, TSRANGE, TSMULTIRANGE, TSTZMULTIRANGE, TSQUERY -> CodeBlock.of(
         "bindObject(%L, %L, %M)\n",
         columnIndex,
         value,
@@ -69,7 +70,7 @@ enum class PostgreSqlType(override val javaType: TypeName) : DialectType {
         BIG_INT -> "$cursorName.getLong($columnIndex)"
         DATE, TIME, TIMESTAMP, TIMESTAMP_TIMEZONE, UUID -> "$cursorName.getObject<%T>($columnIndex)"
         NUMERIC -> "$cursorName.getBigDecimal($columnIndex)"
-        ENUM, INTERVAL, JSON, TSVECTOR, TSTZRANGE, TSRANGE, TSMULTIRANGE, TSTZMULTIRANGE, XML, TSQUERY, GEOMETRY, GEOGRAPHY -> "$cursorName.getString($columnIndex)"
+        LTREE, ENUM, INTERVAL, JSON, TSVECTOR, TSTZRANGE, TSRANGE, TSMULTIRANGE, TSTZMULTIRANGE, XML, TSQUERY, GEOMETRY, GEOGRAPHY -> "$cursorName.getString($columnIndex)"
       },
       javaType,
     )